fix(display):modify set resolution when marge mode
Fix:7278
当合并模式下设置刷新率时，请求设置刷新率的信号第一个参数为nullptr
且保有width,height,rate 三个数据，
module会将所有屏幕设置为满足三个条件的分辨率

+++ src/frame/window/modules/display/customsettingdialog.h
@@ -79,7 +79,7 @@ struct ResolutionDate {
     qint32 id;
     qint32 w;
     qint32 h;
-    qint32 rate{0};
+    double rate{0.0};
 };
 
 Q_SIGNALS:
